Split `test_codepage` into `test_stream_io` & `test_encoding`
Especially to create a test file to test the `encoding.hpp` header (i.e. the `to_utf`, `from_utf` functions)
